Added filters for TodoItems, adjusted Show view for Owners and implemented filters

+++ HouseholdManagementSystem/Controllers/OwnerDataController.cs
@@ -128,10 +128,16 @@ namespace HouseholdManagementSystem.Controllers
                 return NotFound();
             }
 
+            bool hasAssignedItems = db.TodoItems.Any(t=> t.AssignedToOwnerId == owner.OwnerId);
+            if (hasAssignedItems)
+            {
+                return Content(HttpStatusCode.Conflict, "You cannot delete this owner. This owner has assigned tasks");
+            }
+
             db.Owners.Remove(owner);
             db.SaveChanges();
 
-            return Ok();
+            return Ok(owner);
         }
